Add Goetze File Server RFC
Authored in Khondker's Computer Networks class in 2008,
by Ryan Lewis and I. His version is also on GitHub, and
I'm glad he uploaded the thing because it reminded me of
its existence.

His is at the following URL:
https://github.com/cosmotron/class/blob/master/cs455/rfc/rfc.txt

Inspiration for this file server came from the small but
delicious caramel cream candy made by the Goetze's Candy
company. We swear.